Add email reminder functionality
- Both backend and frontend code for the functionality has been written.
- Currently using node-cron for checking db every minute.
- Nodemailer is being used in this version to send emails,
this can later be changed to some third-party service if need be.
- The code has been tested properly via the frontend, however,
backend testing (via Postman) was not so rigorous before this commit.
- In this version, the user will se a "Delete Reminder" button,
if a reminder has already been set for the card,
and a "Set Reminder" button otherwise. (This means that this version
does not allow more than one reminder for a particular card)